+++ .ansible/collections/ansible_collections/netapp/ontap/plugins/modules/na_ontap_volume_autosize.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2019-2025, NetApp, Inc
+# (c) 2019-2026, NetApp, Inc
 # GNU General Public License v3.0+
 # (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
@@ -90,6 +90,31 @@ options:
     - The default shrink theshold is 50%. It is an error for the shrink threshold to be greater than or equal to the grow threshold.
     - Range between 0 and 100
     type: int
+
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.4.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
+notes:
+  - Supports AWS Lambda proxy functionality when using REST. See the README file for examples.
 '''
 
 EXAMPLES = """
@@ -152,6 +177,9 @@ class NetAppOntapVolumeAutosize:
             reset=dict(required=False, type='bool'),
             shrink_threshold_percent=dict(required=False, type='int')
         ))
+
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
+
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
             supports_check_mode=True,
@@ -162,7 +190,8 @@ class NetAppOntapVolumeAutosize:
                 ['reset', 'grow_threshold_percent'],
                 ['reset', 'shrink_threshold_percent'],
                 ['reset', 'mode']
-            ]
+            ],
+            required_if=[['use_lambda', True, ('lambda_config',)]],
         )
         self.na_helper = NetAppModule()
         self.parameters = self.na_helper.set_parameters(self.module.params)
@@ -176,6 +205,8 @@ class NetAppOntapVolumeAutosize:
             if self.parameters.get('reset'):
                 self.module.fail_json(msg="Rest API does not support reset, please switch to ZAPI")
         else:
+            if self.parameters.get('use_lambda'):
+                self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
             if not netapp_utils.has_netapp_lib():
                 self.module.fail_json(msg=netapp_utils.netapp_lib_is_required())
             self.server = netapp_utils.setup_na_ontap_zapi(module=self.module, vserver=self.parameters['vserver'])